[FEATURE] Add paypal entities
