+++ src/entities/payment/paypal/access-token.entity.ts
@@ -1,0 +1,3 @@
+export interface AccessToken {
+  access_token: string;
+}
