+++ src/gateways/payment/paypal/services/paypal.service.ts
@@ -1,71 +1,9 @@
 import { HttpService } from '@nestjs/axios';
 import { Injectable } from '@nestjs/common';
 
-interface Order {
-  id: string;
-  intent: string;
-  status: string;
-  purchase_units: PurchaseUnit[];
-  payer: Payer;
-  links: Link[];
-}
-
-interface PurchaseUnit {
-  amount: Amount;
-  description: string;
-  items: Item[];
-}
-
-interface Amount {
-  currency_code: string;
-  total: number;
-}
-
-interface Item {
-  name: string;
-  sku: string;
-  quantity: number;
-  price: number;
-}
-
-interface Payer {
-  name: string;
-  email: string;
-}
-
-interface Link {
-  rel: string;
-  href: string;
-}
-
-// ----
-interface OrderCreateRequest {
-  intent: 'CAPTURE';
-  purchase_units: PurchaseUnit[];
-}
-
-interface PurchaseUnit {
-  amount: Amount;
-  description: string;
-  items: Item[];
-}
-
-interface Amount {
-  currency_code: string;
-  total: number;
-}
-
-interface Item {
-  name: string;
-  sku: string;
-  quantity: number;
-  price: number;
-}
-
-// ---
-interface AccessToken {
-  access_token: string;
-}
+import { AccessToken } from '../../../../entities/payment/paypal/access-token.entity';
+import { OrderCreateRequest } from '../../../../entities/payment/paypal/create-order.entity';
+import { Order } from '../../../../entities/payment/paypal/order.entity';
 
 @Injectable()
 export class PaypalGateway {
